Exam Preparation Three

+++ 14.Exam Preparation Three/CommandInterpreter/CommandInterpreter.cs
@@ -1,26 +1,21 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CommandInterpreter
+﻿namespace CommandInterpreter
 {
+    using System;
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Text;
+    using System.Threading.Tasks;
+
     class CommandInterpreter
     {
         static void Main(string[] args)
         {
-            List<string> nums = Console.ReadLine().Split( new char[] {' '}, StringSplitOptions.RemoveEmptyEntries).ToList();
-                    
+            List<string> nums = Console.ReadLine().Split(new char[] { ' ' }, StringSplitOptions.RemoveEmptyEntries).ToList();
 
-            
             string[] command = Console.ReadLine().Split().ToArray();
 
             while (!command[0].Equals("end"))
             {
-
-                
-
                 List<string> newList = new List<string>();
 
                 switch (command[0])
@@ -30,30 +25,6 @@ namespace CommandInterpreter
                         int start = int.Parse(command[2]);
                         int end = int.Parse(command[4]);
 
-                        // proverqvame za nevalidni ili negativni elementi
-                        if (start < 0  
-                            || start >= nums.Count
-                            || end < 0 
-                            || (start + end) > nums.Count)
-                        {
-                            Console.WriteLine("Invalid input parameters.");
-                            break;
-                        }
-
-                        newList = nums
-                              .Skip(start)    // skopvame vsichko do start
-                              .Take(end)      // vzimame vsichko do end
-                              .Reverse()     // reversvame
-                              .ToList();
-
-                        nums.RemoveRange(start, end);
-                        nums.InsertRange(start, newList);
-                        break;
-                    case "sort":
-                        start = int.Parse(command[2]);      // VUV SWITCH NQMA NUJDA PAK DA INICIALIZIRAME PROMENLIVITE, ZA TOVA NQMA INT OTPRED
-                        end = int.Parse(command[4]);
-
-                        // proverqvame za nevalidni ili negativni elementi
                         if (start < 0
                             || start >= nums.Count
                             || end < 0
@@ -64,12 +35,34 @@ namespace CommandInterpreter
                         }
 
                         newList = nums
-                             .Skip(start)    // skopvame vsichko do start
-                             .Take(end)      // vzimame vsichko do end
-                             .OrderBy(e => e)   // TAKA GO SORTIRAME             
+                              .Skip(start)
+                              .Take(end)
+                              .Reverse()
+                              .ToList();
+
+                        nums.RemoveRange(start, end);
+                        nums.InsertRange(start, newList);
+                        break;
+                    case "sort":
+                        start = int.Parse(command[2]);
+                        end = int.Parse(command[4]);
+
+                        if (start < 0
+                            || start >= nums.Count
+                            || end < 0
+                            || (start + end) > nums.Count)
+                        {
+                            Console.WriteLine("Invalid input parameters.");
+                            break;
+                        }
+
+                        newList = nums
+                             .Skip(start)
+                             .Take(end)
+                             .OrderBy(e => e)
                              .ToList();
 
-                        
+
                         nums.RemoveRange(start, end);
                         nums.InsertRange(start, newList);
 
@@ -77,13 +70,13 @@ namespace CommandInterpreter
                     case "rollLeft":
                         int count = int.Parse(command[1]);
 
-                        if(count < 0)
+                        if (count < 0)
                         {
                             Console.WriteLine("Invalid input parameters.");
                             break;
                         }
 
-                        for (int i = 0; i < (count % nums.Count); i++)  // MAHAME PURVIQ I GO SLAGAME NAKRAQ
+                        for (int i = 0; i < (count % nums.Count); i++)
                         {
                             string element = nums.First();
                             nums.Remove(nums.First());
@@ -100,27 +93,24 @@ namespace CommandInterpreter
                             break;
                         }
 
-                        for (int i = 0; i < (count % nums.Count); i++)  // MAHAME POSLEDNIQ I GO SLAGAME NAKRAQ
+                        for (int i = 0; i < (count % nums.Count); i++)
                         {
-                           string element = nums.Last();
+                            string element = nums.Last();
                             nums.Remove(nums.Last());
                             nums.Insert(0, element);
                         }
                         break;
-                    default:                      
+                    default:
                         break;
                 }
 
-
                 command = Console.ReadLine().Split();
-
             }
-                
+
 
             string output = string.Join(", ", nums);
-            Console.WriteLine($"[{output}]");               // printirame si nums
 
-
+            Console.WriteLine($"[{output}]");
         }
     }
 }
